node2vec data reader (#1584)
* Implement node2vec data reader and distributed embedding layer

* Squashed commit of the following:

commit 76f2fb662206984b2c3d09ef9945558872f402d3
Author: Tim Moon <moon13@llnl.gov>
Date:   Tue Jun 30 16:31:02 2020 -0700

    Make sure to use mutable weights buffer in dist embedding layer

commit 9ee1eba63b299614a298a53d3865e0ce124d9af0
Merge: f13900870d a9e92a91c1
Author: Tim Moon <moon13@llnl.gov>
Date:   Tue Jun 30 12:18:44 2020 -0700

    Merge branch 'develop' into dist-embedding-layer2

commit f13900870d3be4d25aad3a6629847441be590ee2
Author: Tim Moon <moon13@llnl.gov>
Date:   Thu Jun 25 16:32:01 2020 -0700

    Throw exception if NVSHMEM is initialized after finalization

commit 0b64f8be60d0b711315d0b820d1556941bd1cdb5
Author: Tim Moon <moon13@llnl.gov>
Date:   Thu Jun 25 15:11:30 2020 -0700

    Fix hang in GPU dist embedding layer with NVSHMEM 1.0

commit 9466d6e8d7e062d2c276d974f603f8628ad7e1d1
Author: Tim Moon <moon13@llnl.gov>
Date:   Wed Jun 17 13:58:46 2020 -0700

    Expand documentation for dist embedding layer

commit 80d4ed677b650431ba8d5423b0084791d66b3933
Author: Tim Moon <moon13@llnl.gov>
Date:   Wed Jun 17 11:21:28 2020 -0700

    Make vector metadata struct an implementation detail of dist embedding layer

commit ba165b00e8f1c66e6000a5e162a55dff5f787385
Author: Tim Moon <moon13@llnl.gov>
Date:   Mon Jun 15 19:46:31 2020 -0700

    Don't instantiate dist embedding layer if SHMEM/NVSHMEM aren't available

commit f6271c5b220934933be7205184987bf2bf0c913a
Author: Tim Moon <moon13@llnl.gov>
Date:   Fri Jun 12 17:47:44 2020 -0700

    Fix error in dist embedding layer when building without GPU support

commit 6cec5dad95c76d73e92d46d11143db9a359ed9ab
Author: Tim Moon <moon13@llnl.gov>
Date:   Fri Jun 12 15:34:49 2020 -0700

    Disable Bamboo unit tests for dist embedding layer

commit 4b9f39b96ade597ea4fea9b8e99d1fa3b14b574d
Author: Tim Moon <moon13@llnl.gov>
Date:   Fri Jun 12 15:17:54 2020 -0700

    Distributed embedding layer

    Embeddings weights are distributed and are communicated via SHMEM/NVSHMEM. There is an option to perform sparse SGD instead of using the optimizers, which currently only support dense gradients. See https://github.com/timmoon10/lbann/pull/6.

* Move node2vec app to subdir of graph app

* Bugfixes

* Cleaning up node2vec

Update docs. Default behavior is to download graph from SNAP. Restore option to use offline walks. Update HavoqGT commit hash.

* Reformulate node2vec model to consume an entire walk

node2vec data reader outputs an entire walk instead of a context window. node2vec data reader no longer maintains cache of walks.

* Fix bug in node2vec data reader when walker doesn't return expected number of walks

Maintain cache of walks from previous mini-batch steps.

* Replace context windows in node2vec with decaying scaling factors

* Forgot to include a node2vec utility module

* Implement uniform hash layer on GPU

Apply MD5 hash function and rescale to [0,1).

* Implement erf and erfinv layers

* Support random projection embeddings in node2vec app

* Reorganize node2vec implementation

Move random projection and Skip-gram implementations into their own files.

* Update node2vec data reader to use new IO RNG objects

* Add option to launch node2vec as batch job

* Implement random projeciton node2vec with MSE loss

* Support row-major embeddings in node2vec eval script

* Fix compile error in uniform hash layer

* Fix error from merge

* Update release notes

* Revert build script

* Fix compile error in erf/erfinv layers with half type

+++ applications/graph/node2vec/utils/snap.py
@@ -4,51 +4,71 @@ SNAP is the Stanford Network Analysis Platform. See
 https://snap.stanford.edu.
 
 """
+import gzip
 import os
 import os.path
-import urllib.request
-import gzip
 import subprocess
+import urllib.request
+import numpy as np
 
 # Root directory for LBANN graph application
 _root_dir = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
 
 
-def download_graph(name='ego-Facebook',
-                   graph_file=None):
+def download_graph(name='youtube',
+                   graph_file=None,
+                   reorder_indices=True):
     """Download graph edgelist file from SNAP website.
+
+    Returns immediately if graph has already been downloaded.
 
     Args:
         name (str): Name of graph.
         graph_file (str, optional): File where uncompressed edge list
-            will be saved (default: in 'data' directory).
+            will be saved (default: in 'data/graphs' directory).
 
     Returns:
         str: Uncompressed edge list file.
 
     """
 
-    # Graphs from SNAP
+    # Return immediately if graph is already downloaded
+    if not graph_file:
+        graph_file = os.path.join(_root_dir, 'data', 'graphs', name, 'graph.txt')
+    if os.path.exists(graph_file):
+        return graph_file
+    graph_file = os.path.realpath(graph_file)
+
+    # Graphs URLs
     download_urls = {
-        'ego-Facebook': 'http://snap.stanford.edu/data/facebook_combined.txt.gz',
+        'facebook': 'http://snap.stanford.edu/data/facebook_combined.txt.gz',
+        'youtube': 'https://snap.stanford.edu/data/bigdata/communities/com-youtube.ungraph.txt.gz',
+        'friendster': 'https://snap.stanford.edu/data/bigdata/communities/com-friendster.ungraph.txt.gz',
     }
 
     # Paths
-    if not graph_file:
-        graph_file = os.path.join(_root_dir, 'data', name, 'graph.txt')
     data_dir = os.path.dirname(graph_file)
     if not os.path.isdir(data_dir):
         os.makedirs(data_dir)
     data_dir = os.path.realpath(data_dir)
-    graph_file = os.path.realpath(graph_file)
     compressed_file = graph_file + '.gz'
 
     # Download and uncompress graph file
     urllib.request.urlretrieve(download_urls[name],
                                filename=compressed_file)
-    with gzip.open(compressed_file, 'rb') as in_file:
-        with open(graph_file, 'wb') as out_file:
-            out_file.write(in_file.read())
+    if reorder_indices:
+        edges = np.loadtxt(compressed_file, dtype=np.uint64)
+        index_map = {}
+        for v in np.nditer(edges):
+            v = v[()]
+            if v not in index_map:
+                index_map[v] = np.uint64(len(index_map))
+        edges = np.vectorize(index_map.get)(edges)
+        np.savetxt(graph_file, edges, fmt='%u', delimiter='\t')
+    else:
+        with gzip.open(compressed_file, 'rb') as in_file:
+            with open(graph_file, 'wb') as out_file:
+                out_file.write(in_file.read())
 
     return graph_file
